# utils/Swagger/oas.yaml
openapi: 3.0.3
servers:
  - url: 'https://test.es.sign.fiskaly.com/api/v1'
    description: |
      Entorno de prueba (Test)
  - url: 'https://live.es.sign.fiskaly.com/api/v1'
    description: |
      Entorno de producción (Live)
info:
  version: '1.13.0'
  title: fiskaly SIGN ES
  contact:
    email: 'info@fiskaly.com'
    name: 'Contacto'
    url: 'https://www.fiskaly.com/es'
  description: |
    Información Legal: [fiskaly.com/impressum](https://www.fiskaly.com/es/aviso-legal)
    <br>
    Política de Privacidad: [fiskaly.com/datenschutz](https://www.fiskaly.com/es/privacy-policy)
    <br>
    Solución: [fiskaly.com/signes](https://www.fiskaly.com/es/signes)
    <br>
    Guía para Desarrolladores: [developer.fiskaly.com/es/sign-es](https://developer.fiskaly.com/es/sign-es/introduccion)

    # Introducción

    **SIGN ES**, la [solución](https://www.fiskaly.com/es/signes) de fiskaly, ofrece una [RESTful](https://es.wikipedia.org/wiki/Transferencia_de_Estado_Representacional) API que asegura el cumplimiento con las normativas fiscales españolas como **TicketBAI** y **Verifactu**.

    **SIGN ES** proporciona un entorno de prueba para fines de integración y pruebas, así como un entorno de producción dedicado para realizar el envío de transacciones oficiales (facturas). Para acceder al entorno de producción, [contacta con nuestro equipo de ventas](mailto:sales@fiskaly.com).

...

    - tiene URLs orientadas a recursos,
...
    - usa verbos y códigos de estado HTTP estándar, y  
...

...

...

...

...

...

...
paths:
...
  /taxpayer:
    put:
      summary: Crear contribuyente
      description: Este endpoint se utiliza para crear la información de un contribuyente (emisor de facturas)
      tags:
        - 'Contribuyente'
      security:
        - JWT: []
      operationId: createTaxpayer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaxpayerCreateRequest'
      responses:
        '200':
          $ref: '#/components/responses/TaxpayerResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '405':
          $ref: '#/components/responses/MethodNotAllowedResponse'
        '409':
          $ref: '#/components/responses/ResourceConflictResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    patch:
      summary: Actualizar contribuyente
      description: 'Para actualizar el estado de un contribuyente específico, este endpoint permite desactivar un contribuyente de forma permanente y/o actualizar los metadatos del contribuyente.'
      tags:
        - 'Contribuyente'
      security:
        - JWT: []
      operationId: updateTaxpayer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: object
                  properties:
                    state:
                      $ref: '#/components/schemas/TaxpayerState'
                  additionalProperties: false
                metadata:
                  $ref: '#/components/schemas/Metadata'
              additionalProperties: false
      responses:
        '200':
          $ref: '#/components/responses/TaxpayerResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '404':
          $ref: '#/components/responses/ResourceNotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    get:
      summary: Obtener contribuyente
      description: Este endpoint se utiliza para obtener la información del contribuyente (emisor de facturas).
      tags:
        - 'Contribuyente'
      security:
        - JWT: []
      operationId: retrieveTaxpayer
      responses:
        '200':
          $ref: '#/components/responses/TaxpayerResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '404':
          $ref: '#/components/responses/ResourceNotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  '/signers/{signer_id}':
    put:
      summary: Crear dispositivo firmante
      description: |-
        Este endpoint se utiliza para crear un dispositivo firmante. **SIGN ES** gestiona los certificados para Verifactu y TicketBAI.

        Debe existir al menos un dispositivo de firma para cada organización gestionada para poder emitir facturas.  
        Un dispositivo de firma permite utilizar un certificado de dispositivo exclusivo para TicketBAI del tipo "Punto de facturación" y mantiene el encadenamiento de facturas para uno o varios dispositivos de cliente.

        Para TicketBAI, el certificado firmante emplea un certificado de tipo "Punto de facturación", para la firma y envio de registros encadenados. Cada dispositivo de firma requiere un certificado de dispositivo. Si no se provee un certificado externo durante la creación del dispositivo, **SIGN ES** asignara un certificado gestionado por fiskaly GmbH.

        Para Verifactu, **SIGN ES** utilizara por defecto un certificado electronico gestionado por fiskaly GmbH, que se empleara para la firma y envio a la agencia tributaria de registros encadenados.

        Esta llamada es idempotente.
      tags:
        - 'Firmantes'
      security:
        - JWT: []
      operationId: createSigner
      parameters:
        - $ref: '#/components/parameters/SignerIdentifier'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  $ref: '#/components/schemas/Signer'
                metadata:
                  $ref: '#/components/schemas/Metadata'
              additionalProperties: false
      responses:
        '200':
          $ref: '#/components/responses/SignerResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '409':
          $ref: '#/components/responses/ResourceConflictResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    patch:
      summary: Actualizar dispositivo firmante
      description: 'Para actualizar el estado de un dispositivo firmante específico, este endpoint permite deshabilitar (desactivar) un dispositivo de firma de forma permanente y/o actualizar los metadatos del dispositivo firmante.'
      tags:
        - 'Firmantes'
      security:
        - JWT: []
      operationId: updateSigner
      parameters:
        - $ref: '#/components/parameters/SignerIdentifier'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: object
                  properties:
                    state:
                      $ref: '#/components/schemas/DeviceState'
                  additionalProperties: false
                metadata:
                  $ref: '#/components/schemas/Metadata'
              additionalProperties: false
      responses:
        '200':
          $ref: '#/components/responses/SignerResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '404':
          $ref: '#/components/responses/ResourceNotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    get:
      summary: Obtener un firmante
      description: Obtiene el dispositivo firmante específico de una organización.
      tags:
        - 'Firmantes'
      security:
        - JWT: []
      operationId: retrieveSigner
      parameters:
        - $ref: '#/components/parameters/SignerIdentifier'
      responses:
        '200':
          $ref: '#/components/responses/SignerResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '404':
          $ref: '#/components/responses/ResourceNotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /signers:
    get:
      summary: Listar dispositivos firmantes
      description: Obtiene una lista de dispositivos firmantes de una organización.
      tags:
        - 'Firmantes'
      security:
        - JWT: []
      operationId: listSigners
      parameters:
        - $ref: '#/components/parameters/PaginationLimitQuery'
        - $ref: '#/components/parameters/PaginationTokenQuery'
      responses:
        '200':
          $ref: '#/components/responses/SignersResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
components:
  parameters:
    SignerIdentifier:
      description: 'Identifica un firmante (dispositivo de firma) con un identificador aleatorio en formato [UUIDv4](#section/Producto/Recursos).'
      in: path
      name: signer_id
      required: true
      schema:
        $ref: '#/components/schemas/UniversallyUniqueIdentifierV4'
...
    PaginationLimitQuery:
      in: query
      name: limit
      required: false
      schema:
        $ref: '#/components/schemas/PaginationLimit'
    PaginationTokenQuery:
      in: query
      name: token
      required: false
      schema:
        $ref: '#/components/schemas/PaginationToken'
  responses:
...
    TaxpayerResponse:
      description: Devuelve la información del contribuyente.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TaxpayerResource'
    SignersResponse:
      description: Devuelve una lista de dispositivos de firma.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SignerResources'
    SignerResponse:
      description: Devuelve un dispositivo de firma.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SignerResource'
...
    BadRequestResponse:
      description: Solicitud incorrecta (no válida)
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
    UnauthorizedAccessResponse:
      description: Acceso no autorizado.
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
    ResourceNotFoundResponse:
      description: El recurso solicitado no existe.
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
    ResourceConflictResponse:
      description: El recurso solicitado tiene un conflicto.
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
    MethodNotAllowedResponse:
      description: The request method is not allowed.
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
    InternalServerErrorResponse:
      description: Ocurrió un error interno del servidor
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
...
  schemas:
    ErrorResponse:
      description: Representación de respuestas de error.
      type: object
      required:
        - code
        - message
        - status
        - error
      properties:
        status:
          $ref: '#/components/schemas/ErrorStatus'
        code:
          $ref: '#/components/schemas/ErrorCode'
        error:
          $ref: '#/components/schemas/ErrorText'
        message:
          $ref: '#/components/schemas/ErrorMessage'
      additionalProperties: false
    ErrorStatus:
      description: Código de estado de error HTTP basado en números enteros.
      type: integer
      format: int32
      example: 401
...
    TaxpayerCreateRequest:
      type: object
      required:
        - content
      properties:
        content:
          $ref: '#/components/schemas/TaxpayerCreateContent'
        metadata:
          $ref: '#/components/schemas/Metadata'
      additionalProperties: false
    TaxpayerResource:
      description: Representación de un recurso de contribuyente.
      type: object
      required:
        - content
      properties:
        content:
          $ref: '#/components/schemas/Taxpayer'
        metadata:
          $ref: '#/components/schemas/Metadata'
      additionalProperties: false
    TaxpayerCreateContent:
      description: 'Representa la información del contribuyente que se utiliza para emitir facturas para todos los dispositivos de firma dentro del contexto de la organización gestionada, asociada al [JWT](#section/JWT). Consta de información del emisor `issuer` de la factura y del territorio `territory` que representa el domicilio fiscal del contribuyente.'
      type: object
      required:
        - issuer
        - territory
      properties:
        issuer:
          $ref: '#/components/schemas/NationalIdentification'
        territory:
          $ref: '#/components/schemas/Territory'
      additionalProperties: false
...
    TaxpayerState:
      description: 'Un contribuyente en el sistema se establecerá en el estado `ENABLED` (HABILITADO) y solo se puede establecer como `DISABLED` (DESHABILITADO), pero no al revés.'
      type: string
      enum:
        - ENABLED
        - DISABLED
    Signer:
      description: Representa el dispositivo de firma que realiza las operaciones de encadenamiento y firma de facturas.
      type: object
      required:
        - certificate
      properties:
        certificate:
          $ref: '#/components/schemas/Certificate'
      additionalProperties: false
    SignerResources:
      description: Representación de recursos de firma.
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SignerResource'
        pagination:
          $ref: '#/components/schemas/Pagination'
      additionalProperties: false
    SignerResource:
      description: Representación de un recurso de firma.
      type: object
      required:
        - content
      properties:
        content:
          $ref: '#/components/schemas/SigningDevice'
        metadata:
          $ref: '#/components/schemas/Metadata'
      additionalProperties: false
...
    NationalIdentification:
      description: Representa información sobre un emisor nacional (nombre legal y número de identificación fiscal).
      type: object
      required:
        - legal_name
        - tax_number
      properties:
        legal_name:
          $ref: '#/components/schemas/LegalName'
        tax_number:
          $ref: '#/components/schemas/TaxNumber'
      additionalProperties: false
...
    Certificate:
      description: 'Representación del par de claves criptográficas en formato PKCS12 o PEM. Por defecto, el dispositivo de firma utiliza un certificado de dispositivo gestionado por fiskaly. Es posible proporcionar un certificado de dispositivo externo opcional acorde a las instrucciones en la sección de la guía [Certificado de dispositivo opcional](/es/sign-es/optional_device_certificate). Un solo certificado puede incluirse en cada firmante.'
      discriminator:
        propertyName: type
        mapping:
          P12: '#/components/schemas/CertificateP12'
          PEM: '#/components/schemas/CertificatePEM'
      oneOf:
        - $ref: '#/components/schemas/CertificateP12'
        - $ref: '#/components/schemas/CertificatePEM'
...
    Territory:
      description: |
        El territorio del domicilio fiscal del contribuyente. Esto indicará si se aplica la normativa TicketBAI o Verifactu. Actualmente, SIGN ES admite las siguientes regiones territoriales:
          - `"ARABA"`          `"Araba/Álava (País Vasco)"`
          - `"BIZKAIA"`        `"Bizkaia/Vizcaya (País Vasco)"`
          - `"GIPUZKOA"`       `"Gipuzkoa/Guipúzcoa (País Vasco)"`
          - `"NAVARRE"`        `"Nafarroa/Navarra"`
          - `"CANARY_ISLANDS"` `"Islas Canarias"`
          - `"CEUTA"`          `"Ceuta"`
          - `"MELILLA"`        `"Melilla"`
          - `"SPAIN OTHER"`    `"Otro territorio nacional"`
      type: string
      enum:
        - ARABA
        - BIZKAIA
        - GIPUZKOA
        - NAVARRE
        - CANARY_ISLANDS
        - CEUTA
        - MELILLA
        - SPAIN_OTHER
    DeviceState:
      description: 'Cada dispositivo creado en el sistema se establecerá en el estado `ENABLED` (HABILITADO) y solo se puede establecer en `DISABLED` (DESHABILITADO), pero no al revés.'
      type: string
      enum:
        - ENABLED
        - DISABLED
...
    UniversallyUniqueIdentifierV4:
      description: |-
        Un identificador único universal versión 4 (UUIDv4), es una etiqueta empleada para identificar objetos de forma única en
        sistemas informáticos. La representación de UUIDs está definida dentro del [RFC 4122](https://www.rfc-editor.org/rfc/rfc4122).
      type: string
      format: uuid
      pattern: '[a-f0-9]{8}-?[a-f0-9]{4}-?4[a-f0-9]{3}-?[89ab][a-f0-9]{3}-?[a-f0-9]{12}'
      example: 1c81cb86-c2e8-4074-afc3-a0601b2bf063
      x-go-type: uuid.UUID
      x-go-type-import:
        name: uuid
        path: github.com/google/uuid
...
    Metadata:
      description: |-
        Puede emplear este parámetro para adjuntar datos en formato clave/valor al objeto.  
        Los metadatos son útiles para almacenar datos estructurados que complementan el objeto.

        Se pueden adjuntar metadatos en los métodos de actualización de cada recurso. Si se incluye una clave con un valor vacío, se eliminará esa pareja clave/valor.

        Si se adjuntan metadatos a un objeto, estos serán devueltos en las respuestas a peticiones de ese objeto.

        **Nota**: Se pueden especificar hasta 20 claves, cada clave puede contener hasta 40 caracteres, y los valores pueden contener hasta 500 caracteres.
      type: object
      maxProperties: 20
      additionalProperties:
        type: string
        maxLength: 500
      example:
        my_property_1: '1234'
        my_property_2: 'https://my-internal-system/path/to/resource/1234'
...
    PaginationLimit:
      description: Representa el límite de resultados por página para un endpoint de listado.
      type: integer
      minimum: 1
      maximum: 100
      default: 10
    PaginationToken:
      description: 'Representa el token de paginación, que se emplea como puntero hacia la siguiente página de resultados.'
      type: string
      pattern: '^(?:[A-Za-z0-9+/]{4})*(?:[A-Za-z0-9+/]{4}|[A-Za-z0-9+/]{3}=|[A-Za-z0-9+/]{2}={2})$'
      maxLength: 1024
...
  securitySchemes:
    JWT:
      description: |-
        JSON Web Token (JWT) empleado para control de acceso y autorización.
        - Ejemplo de uso: `Bearer <JWT>`
        - Ejemplo de encabezado HTTP: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
      type: http
      scheme: bearer
      bearerFormat: JWT